refactor: split pipeline.rs god-file into modules
