+++ homeboy.json
@@ -726,7 +726,7 @@
         "intra-method-duplication::src/core/observation/store.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/preview_ingress.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/quality.rs::IntraMethodDuplicate",
-        "intra-method-duplication::src/core/rig/pipeline.rs::IntraMethodDuplicate",
+        "intra-method-duplication::src/core/rig/pipeline/requirement_step.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/connection.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/execution/daemon_api.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/execution/extension_parity.rs::IntraMethodDuplicate",
@@ -797,8 +797,6 @@
         "structural::src/core/preview_ingress.rs::HighItemCount",
         "structural::src/core/refactor/rename/mod.rs::GodFile",
         "structural::src/core/release/plan_steps.rs::GodFile",
-        "structural::src/core/rig/pipeline.rs::GodFile",
-        "structural::src/core/rig/pipeline.rs::HighItemCount",
         "structural::src/core/runner/connection.rs::HighItemCount",
         "structural::src/core/runner/evidence.rs::HighItemCount",
         "structural::src/core/runner/lab/agent_task_bridge.rs::HighItemCount",
